requirements.txt

Django>=6.0.2,<7.0
djangorestframework>=3.15.0
django-filter>=24.0
drf-yasg>=1.21.7


python-dotenv>=1.0.0


pytest>=8.0.0
pytest-django>=4.8.0
coverage>=7.0.0


